pyproject.toml
[project]
name = "pai"
version = "0.1.0"
description = "LLM-driven CLI"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.3",
  "pyyaml>=6.0.1",
  "python-dotenv>=1.0.1",
  "openai>=1.40.0",
]

[project.scripts]
pai = "pai.cli:run"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
